# Improve performance of `brew readall`