* Significantly improve base performance (60s -> 30s for Homebrew/core before the extra checks were added)
* Add `to_hash` check to ensure the method used for API generation (and `brew info` etc) doesn't error. This increases that 30s back up to 47s (after some optimisations) _but_ does mean we could remove the 90 second `brew generate-formula-api` call in tap_syntax checks as the other elements of that command shouldn't be affected by formula changes.

These changes affect formulae only. I have not touched casks.