@apainintheneck 2023-09-15
If we're going to move the `Formula#to_h` check, I think that it'd make more sense to have it in `brew audit` than in `brew readall`.