@Bo98 2023-09-15
We ideally want to `to_hash` for all OS and arch combos if we want to replace `brew generate-formula-api`, but `brew audit` (in our CI at least) only does the running OS and would probably add ~45s to `brew audit`'s runtime to start doing that, rather than ~15s to `brew readall`.

However we could just get rid of `brew readall` entirely in CI as adding `to_hash` to `brew audit` means that we have already done all the formula loading checks `brew readall` would normally do.