@Bo98 2023-09-15
In theory `to_hash` should only break if there is fundamentally wrong with the formula, such as passing garbage somewhere. Now that Sorbet is enabled, it probably shouldn't ever happen now to be honest but I'm not sure if that's guaranteed.