@apainintheneck 2023-09-15
Sorry, where would we be removing the `brew generate-api-*` check from in CI: here or in core? To be pedantic I believe we use `#to_hash_with_variations` when generating the API. This probably gets us most of the way there though and I don't think we've had any problems with generation on either the formula or the cask side recently. The only bugs I remember happening recently were with reading casks.

I still think it's a bit of a weird thing to add to `brew readall` but it sounds like it will save a minute in CI, is that right?