@MikeMcQuaid 2023-09-15
> However we could just get rid of `brew readall` entirely in CI as adding `to_hash` to `brew audit` means that we have already done all the formula loading checks `brew readall` would normally do in order to be able to call `to_hash` on a loaded formula object on all simulated OS and arch combos. The `to_hash` is one line of code addition here.

This makes sense to me. I was also leaning towards "let's just have a single check that loads all formulae". From that perspective, I agree `to_h` makes more sense in `audit` than `readall`. 

> I still think it's a bit of a weird thing to add to `brew readall` but it sounds like it will save a minute in CI, is that right?

Sounds like it.

> Sorry, where would we be removing the brew generate-api-* check from in CI: here or in core?

Core, not here.