@apainintheneck 2023-09-16
I've changed my mind. I think adding the `Formula#to_hash` check here makes sense to me.