@Bo98 2023-09-28
Splits this into:

* `brew readall` performance improvements: this PR
* `Formula#to_hash` performance improvements: #16052
* Adding `to_hash` to `brew readall`: https://github.com/Bo98/brew/commit/f58168afc823399581794864c7f9aaed797f1b27 (requires the previous two PRs).